requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
PyYAML==6.0.2
-e git+ssh://git@github.com/mattiasblixt/settings.git@26d20671c6a80c130abcc769f52446a1e905432d#egg=settings_project
tomli==2.1.0
